move cosmos serilaizer to utils
